# Improve marching cubes performance on GPUs