--- a/src/axom/primal/geometry/Vector.hpp
+++ b/src/axom/primal/geometry/Vector.hpp
@@ -126,12 +126,30 @@ AXOM_HOST_DEVICE Vector<T, NDIMS> operator*(const T scalar,
 /*!
  * \brief Scalar division of vector; Scalar on rhs.
  * \param [in] vec vector instance
- * \param [in]n scalar user-supplied scalar.
+ * \param [in] scalar user-supplied scalar.
  * \return C resulting vector, \f$ \ni: C_i = vec_i / scalar, \forall i\f$
  * \pre scalar != 0.0
  */
 template <typename T, int NDIMS>
 Vector<T, NDIMS> operator/(const Vector<T, NDIMS>& vec, const T scalar);
+
+/*!
+ * \brief Element-wise < operator.
+ * \param [in] vec vector instance
+ * \param [in] scalar user-supplied scalar.
+ * \return Whether all vector elements are < a scalar.
+ */
+template <typename T, int NDIMS>
+bool operator<(const Vector<T, NDIMS>& vec, const T scalar);
+
+/*!
+ * \brief Element-wise >= operator.
+ * \param [in] vec vector instance
+ * \param [in] scalar user-supplied scalar.
+ * \return Whether all vector elements are >= a scalar.
+ */
+template <typename T, int NDIMS>
+bool operator>=(const Vector<T, NDIMS>& vec, const T scalar);
 
 /*!
  * \brief Overloaded output operator for vectors
@@ -696,6 +714,30 @@ std::ostream& operator<<(std::ostream& os, const Vector<T, NDIMS>& vec)
 {
   vec.print(os);
   return os;
+}
+
+//------------------------------------------------------------------------------
+template <typename T, int NDIMS>
+inline bool operator<(const Vector<T, NDIMS>& vec, const T scalar)
+{
+  bool result(true);
+  for(int d = 0; d < NDIMS; ++d)
+  {
+    result &= vec[d] < scalar;
+  }
+  return result;
+}
+
+//------------------------------------------------------------------------------
+template <typename T, int NDIMS>
+inline bool operator>=(const Vector<T, NDIMS>& vec, const T scalar)
+{
+  bool result(true);
+  for(int d = 0; d < NDIMS; ++d)
+  {
+    result &= vec[d] >= scalar;
+  }
+  return result;
 }
 
 //------------------------------------------------------------------------------