--- a/src/axom/quest/ArrayIndexer.hpp
+++ b/src/axom/quest/ArrayIndexer.hpp
@@ -283,7 +283,7 @@ public:
     return m_strides;
   }
 
-  //!@brief Whether a vector is a permutation vector
+  //!@brief Whether a StackArray represents a permutation.
   bool isPermutation(const axom::StackArray<std::uint16_t, DIM>& v)
   {
     // v is a permutation if all its values are unique and in [0, DIM).