# Improve marching cubes performance on GPUs

- This PR is a performance improvement, with some re-factoring
- It does the following:
  - Implements a fully data-parallel scan stage for marching cubes, without a parallel sort (which was too slow).
  - Reuse memory between calls to `computeIsocontour` to save time, because `cudaMalloc` and `cudaFree` are slow.
  - Implement a `clear` function to allow re-using the `MarchingCubes` object, saving the cost of destructing, constructing and re-allocating memory.
  - Share scratch spaces among domain when possible to reduce memory use.
  - Re-combines the two implementations (full and hybrid parallel) back into the same class.
  - Provides more efficient access to output data, keeping data in device memory.
  - Set up multidimensional array scratch space with stride ordering that matches application data's, and nest the loops by the stride order (to minimize cache misses).

Performance improvements are demonstrated on the following problem:
 - 4M cells in a 159x159x159 mesh.
 - Domain is a cube with corners at (0,0,0) and (2,2,2)
 - Mesh is perfectly load balanced in 4 domains (for rzansel run) or 36 domains (for rzgenie run).
   rzansel was run with 4 CPUs with 1 GPU each.  rzgenie was run with 36 CPUs.
 - Contour surface is a sphere with radius 1.25, centered at (0,0,0)
 - The same contour is repeated 100 times.  After each 10 reps, the mesh is replaced (albeit with the same test mesh) to simulate a change in mesh metadata.

The following 2 charts show the profiler outputs for part of the runs.  The first chart is before memory reuse, the second is after.  Green boxes represent instrumented functions, the top row being the `computeIsocontour` calls.  Blank areas in this row include the constructor, destructor, set-up and tear-down, which are significant before optimization but practically gone after.  The red boxes represent CUDA mallocs, the pink, CUDA frees.  In the first `computeIsocontour` call, the optimized code is moderately faster than the base code.  In the steady state afterward, the optimized code is much faster, even across mesh changes.  In the roughly 40ms shown, the base code completes 2 contours, and the optimized code completes about 50.

Base code profile:
![rzansel profile base](https://github.com/LLNL/axom/assets/45609916/2eefa7dc-1ca6-420b-8c13-48aaab1141dc)

Optimized code profile:
![rzansel profile optimized](https://github.com/LLNL/axom/assets/45609916/ff8133e8-0259-498d-838f-8397ce73bbd3)

Times in the following table include memory allocation and repetitions of `computeIsocontour`.  Units are seconds.  The first row, `repsTimer`, shows values for a timer in the benchmark code.  All other rows are from the NVIDIA profiling tool.

   | timer        | genie base | genie new | ansel base | ansel new | accel |
   |--------------|------------|-----------|------------|-----------|-------|
   | repsTimer    |       0.21 |      0.19 |       0.76 |     0.056 |  13.6 |
   | 100 reps     |            |           |       1.46 |      .080 | 18.3 |
   | first rep    |            |           |       .015 |      .013 |  1.15 |
   | 1st 10 reps  |            |           |        .16 |      .019 |   8.4 |
   | 2nd 10 reps  |            |           |        .16 |     .0066 |  24.2 |
   | last 10 reps |            |           |       .096 |     .0059 |  16.3 |
   | last rep     |            |           |      .0078 |    .00053 |  14.7 |

Timing on the CPU-only rzgenie changes little after optimization, as expected, because this work targets GPU performance.  rzgenie uses the hybrid-parallel implementation, but the numbers verify that the changes do not hinder its performance.

Comparing the base and new results for rzansel shows a 13.6x speed-up according to the benchmark code timer and 18.3x according to the profiler.  The discrepancy in the code timer seems to be caused by running under the profiler.  The timer reports 0.76s when run without the profiler but when run *with* the profiler, it reports 1.48s, much closer than the 1.46 reported by the profiler.  The difference may be due to profiiler overhead.  Other lines show results for various groups of reps.  The base code is relatively consistent, with about 15ms per repetition, although the last rep is only 7.8ms.  The benefit of the optimization is only 1.15x for the first rep, where the new code still has to set up buffers (though less space is required).  Subsequent reps show speed-ups in the range of 8x-24x.